+++ api/handlers/command.go
@@ -2,15 +2,15 @@ package handlers
 
 import (
 	"net/http"
-	
-	"github.com/gin-gonic/gin"
-	"github.com/Bl4omArchie/simple"
-	"github.com/Bl4omArchie/oto/pkg"
+
 	"github.com/Bl4omArchie/oto/models"
+	oto "github.com/Bl4omArchie/oto/pkg"
+	"github.com/Bl4omArchie/simple"
+	"github.com/gin-gonic/gin"
 	"github.com/go-playground/validator"
 )
 
-func CreateCommand(c *gin.Context, cfg *oto.Config) {
+func CreateCommand(c *gin.Context, cfg *oto.Instance) {
 	var cmd models.Command
 
 	if err := c.ShouldBindJSON(&cmd); err != nil {
@@ -32,7 +32,7 @@ func CreateCommand(c *gin.Context, cfg *oto.Config) {
 	c.JSON(http.StatusOK, cmd)
 }
 
-func GetCommands(c *gin.Context, oto *oto.Config) {
+func GetCommands(c *gin.Context, oto *oto.Instance) {
 	cmds, err := simple.GetTable[models.Command](c, oto.Database, -1)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error, couldn't get commands": err.Error()})
@@ -41,7 +41,7 @@ func GetCommands(c *gin.Context, oto *oto.Config) {
 	c.JSON(http.StatusOK, cmds)
 }
 
-func GetCommand(binTag string, cmdName string, c *gin.Context, oto *oto.Config) {
+func GetCommand(execTag string, cmdName string, c *gin.Context, oto *oto.Instance) {
 	cmd, err := simple.GetRowBy[models.Command](c, oto.Database, "name", cmdName)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error, couldn't get command": err.Error()})