+++ api/handlers/job.go
@@ -2,15 +2,15 @@ package handlers
 
 import (
 	"net/http"
-	
-	"github.com/gin-gonic/gin"
-	"github.com/Bl4omArchie/simple"
-	"github.com/Bl4omArchie/oto/pkg"
+
 	"github.com/Bl4omArchie/oto/models"
+	oto "github.com/Bl4omArchie/oto/pkg"
+	"github.com/Bl4omArchie/simple"
+	"github.com/gin-gonic/gin"
 	"github.com/go-playground/validator"
 )
 
-func CreateJob(c *gin.Context, cfg *oto.Config) {
+func CreateJob(c *gin.Context, cfg *oto.Instance) {
 	var job models.Job
 
 	if err := c.ShouldBindJSON(&job); err != nil {
@@ -32,7 +32,7 @@ func CreateJob(c *gin.Context, cfg *oto.Config) {
 	c.JSON(http.StatusOK, job)
 }
 
-func GetJobs(c *gin.Context, oto *oto.Config) {
+func GetJobs(c *gin.Context, oto *oto.Instance) {
 	execs, err := simple.GetTable[models.Job](c, oto.Database, -1)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error, couldn't get commands": err.Error()})
@@ -41,7 +41,7 @@ func GetJobs(c *gin.Context, oto *oto.Config) {
 	c.JSON(http.StatusOK, execs)
 }
 
-func GetJob(jobName string, c *gin.Context, oto *oto.Config) {
+func GetJob(jobName string, c *gin.Context, oto *oto.Instance) {
 	jobCmd, err := simple.GetRowBy[models.Job](c, oto.Database, "name", jobName)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error, couldn't get command": err.Error()})