+++ api/handlers/parameter.go
@@ -1,18 +1,16 @@
 package handlers
 
-
 import (
 	"net/http"
-	
-	"github.com/gin-gonic/gin"
-	"github.com/Bl4omArchie/simple"
-	"github.com/Bl4omArchie/oto/pkg"
+
 	"github.com/Bl4omArchie/oto/models"
+	oto "github.com/Bl4omArchie/oto/pkg"
+	"github.com/Bl4omArchie/simple"
+	"github.com/gin-gonic/gin"
 	"github.com/go-playground/validator"
 )
 
-
-func CreateParameter(c *gin.Context, cfg *oto.Config) {
+func CreateParameter(c *gin.Context, cfg *oto.Instance) {
 	var param models.Parameter
 
 	if err := c.ShouldBindJSON(&param); err != nil {
@@ -34,7 +32,7 @@ func CreateParameter(c *gin.Context, cfg *oto.Config) {
 	c.JSON(http.StatusOK, param)
 }
 
-func GetParameters(c *gin.Context, oto *oto.Config) {
+func GetParameters(c *gin.Context, oto *oto.Instance) {
 	execs, err := simple.GetTable[models.Parameter](c, oto.Database, -1)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error, couldn't get parameters": err.Error()})
@@ -43,8 +41,8 @@ func GetParameters(c *gin.Context, oto *oto.Config) {
 	c.JSON(http.StatusOK, execs)
 }
 
-func GetParameter(binTag string, paramName string, c *gin.Context, oto *oto.Config) {
-	param, err := simple.GetRowBy[models.Parameter](c, oto.Database, binTag, paramName)
+func GetParameter(execTag string, paramName string, c *gin.Context, oto *oto.Instance) {
+	param, err := simple.GetRowBy[models.Parameter](c, oto.Database, execTag, paramName)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error, couldn't get parameter": err.Error()})
 		return
@@ -53,5 +51,5 @@ func GetParameter(binTag string, paramName string, c *gin.Context, oto *oto.Conf
 }
 
 func GetValueTypes(c *gin.Context) {
-    c.JSON(http.StatusOK, []string{"string", "int", "bool"})
+	c.JSON(http.StatusOK, []string{"string", "int", "bool"})
 }