+++ main.go
@@ -4,58 +4,151 @@ import (
 	"fmt"
 	"context"
 
-	"github.com/Bl4omArchie/oto/pkg"
 	"github.com/Bl4omArchie/oto/models"
+	oto "github.com/Bl4omArchie/oto/pkg"
+
+	_ "ariga.io/atlas-provider-gorm/gormschema"
 )
 
-func init_oto(cfg *oto.Instance) error {
+func full_data() error {
 	var ctx context.Context = context.Background()
 
-	err := cfg.AddBinary("nmap", "7.98", "/usr/bin/nmap", "scanning tool")
+	instance, err := oto.NewInstanceOto(".env")
+	if err != nil {
+		return err
+	}
+
+	err = instance.AddExecutable("nmap", "7.98", "/usr/bin/nmap", "scanning tool")
 	if err != nil {
 		fmt.Println(err)
 	}
 
-	s, err := cfg.AddBinarySchema(ctx, "nmap - 7.98")
+	err = instance.AddExecutable("openssl", "3.5.3", "/usr/bin/openssl", "cryptographic tool")
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	err = instance.AddExecutable("masscan", "1.3.9", "/usr/bin/masscan", "scanning tool")
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	s1, err := instance.AddExecutableSchema(ctx, "nmap - 7.98")
 	if err != nil {
 		return err
 	}
+	s2, err := instance.AddExecutableSchema(ctx, "openssl - 3.5.3")
+	if err != nil {
+		return err
+	}
+	s3, err := instance.AddExecutableSchema(ctx, "masscan - 1.3.9")
+	if err != nil {
+		return err
+	}
+
+	err = instance.ImportParameters(ctx, "data/nmap.json", s1)
+	if err != nil {
+		return err
+	}
+
+	err = instance.ImportParameters(ctx, "data/openssl.json", s2)
+	if err != nil {
+		return err
+	}
+
+	err = instance.ImportParameters(ctx, "data/masscan.json", s3)
+	if err != nil {
+		return err
+	}
+
+	err = instance.AddCommand(ctx, "openssl - 3.5.3", "GenRSA", "Generate an rsa keypair", []string{"genpkey", "-algorithm", "-pkeyopt", "-out"}, s2)
+	if err != nil {
+		return err
+	}
+
+	arg1, err := models.FetchParameter(ctx, instance.Database, "flag", "genpkey")
+	arg2, err := models.FetchParameter(ctx, instance.Database, "flag", "-algorithm")
+	arg3, err := models.FetchParameter(ctx, instance.Database, "flag", "-pkeyopt")
+	arg4, err := models.FetchParameter(ctx, instance.Database, "flag", "-out")
 	
-	err = cfg.AddParameter(ctx, "nmap - 7.98", "-sL", "scan option for determine which host are online", false, false, models.String, []string{}, []string{}, s)
+	if err := instance.AddJob(ctx, "GenRSA", "GenRSA-2048", map[*models.Parameter]string{arg1: "", arg2: "RSA", arg3: "rsa_keygen_bits:2048", arg4: "key.pem"}); err != nil {
+		return err
+	}
+
+	out, err := instance.RunJobDemo(ctx, "GenRSA-2048")
 	if err != nil {
 		return err
 	}
 
-	err = cfg.AddParameter(ctx, "nmap - 7.98", "-sT", "scan type", false, false, models.String, []string{"-sL"}, []string{}, s)
+	fmt.Println(out.Stderr, out.Stdout)
+
+	return nil
+}
+
+func launch_demo() error {
+	var ctx context.Context = context.Background()
+
+	instance, err := oto.NewInstanceOto(".env")
 	if err != nil {
 		return err
 	}
 
-	err = cfg.AddParameter(ctx, "nmap - 7.98", "-sK", "scan with -sT", false, false, models.String, []string{}, []string{"-sT"}, s)
-	if err != nil {
-		return err
-	}
-
-	err = cfg.AddParameter(ctx, "nmap - 7.98", "-T", "option depending on -sL", false, false, models.String, []string{"-sL", "-sT"}, []string{}, s)
-	if err != nil {
-		return err
-	}
-
-	err = cfg.AddCommand(ctx, "nmap - 7.98", "-sL", "determine which hosts are online", []string{"-sT", "-T"}, s)
+	err = instance.AddExecutable("openssl", "3.5.3", "/usr/bin/openssl", "cryptographic tool")
 	if err != nil {
 		fmt.Println(err)
 	}
-	// cfg.AddJob(ctx, "nmap - 7.98", "-sL", "target gorm.io", map[string]string{"-sL": "185.199.111.153"})
+
+	s2, err := instance.AddExecutableSchema(ctx, "openssl - 3.5.3")
+	if err != nil {
+		return err
+	}
+	err = instance.ImportParameters(ctx, "data/openssl.json", s2)
+	if err != nil {
+		return err
+	}
+
+	err = instance.AddCommand(ctx, "openssl - 3.5.3", "GenRSA", "Generate an rsa keypair", []string{"genpkey", "-algorithm", "-pkeyopt", "-out"}, s2)
+	if err != nil {
+		return err
+	}
+
+	arg1, err := models.FetchParameter(ctx, instance.Database, "flag", "genpkey")
+	arg2, err := models.FetchParameter(ctx, instance.Database, "flag", "-algorithm")
+	arg3, err := models.FetchParameter(ctx, instance.Database, "flag", "-pkeyopt")
+	arg4, err := models.FetchParameter(ctx, instance.Database, "flag", "-out")
+	
+	if err := instance.AddJob(ctx, "GenRSA", "GenRSA-2048", map[*models.Parameter]string{arg1: "", arg2: "RSA", arg3: "rsa_keygen_bits:2048", arg4: "key.pem"}); err != nil {
+		return err
+	}
+
+	out, err := instance.RunJobDemo(ctx, "GenRSA-2048")
+	if err != nil {
+		return err
+	}
+
+	fmt.Println(out.Stderr, out.Stdout)
+	return nil
+}
+
+func test() error {
+	var ctx context.Context = context.Background()
+
+	instance, err := oto.NewInstanceOto(".env")
+	if err != nil {
+		return err
+	}
+
+	out, err := instance.RunJobDemo(ctx, "GenRSA-2048")
+	if err != nil {
+		return err
+	}
+
+	fmt.Println(out.Stderr, out.Stdout)
 
 	return nil
 }
 
 
 func main() {
-	cfg, err := oto.NewInstanceOto("db/storage.db")
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	init_oto(cfg)
+	fmt.Println(test())
 }